refactor: Split single program file into multiple modules
