+++ pyproject.toml
@@ -3,11 +3,14 @@ name = "app"
 version = "0.1.0"
 description = "Add your description here"
 readme = "README.md"
-requires-python = ">=3.12"
+requires-python = ">=3.11"
 dependencies = [
     "pyinstaller>=6.14.2",
     "pyobjc-framework-corewlan>=11.1",
 ]
+
+[project.scripts]
+SignalSurfer = "src:main:main"
 
 [tool.uv.workspace]
 members = [
